Defensive check in case our destructor is called twice
- I've seen DESTROY called twice before due to some AUTOLOAD-related bugs

+++ lib/Session/Token.pm
@@ -110,7 +110,11 @@ sub get {
 sub DESTROY {
   my ($self) = @_;
 
+  return if !exists $self->{ctx};
+
   _destroy_isaac_context($self->{ctx});
+
+  delete $self->{ctx};
 }
